Fix: On worker shutdown, don't raise error when worker not found
Resolve BE-2046

+++ pkg/worker/worker.go
@@ -441,16 +441,11 @@ func (s *Worker) shutdown() error {
 	// Stops goroutines
 	s.cancel()
 
-	var errs error
-	if worker, err := s.workerRepo.GetWorkerById(s.workerId); err != nil {
-		errs = errors.Join(errs, err)
-	} else if worker != nil {
-		err = s.workerRepo.RemoveWorker(worker)
-		if err != nil {
-			errs = errors.Join(errs, err)
-		}
+	if worker, err := s.workerRepo.GetWorkerById(s.workerId); err == nil && worker != nil {
+		s.workerRepo.RemoveWorker(worker)
 	}
 
+	var errs error
 	err := s.storage.Unmount(s.config.Storage.FilesystemPath)
 	if err != nil {
 		errs = errors.Join(errs, fmt.Errorf("failed to unmount storage: %v", err))
